// src/app/model/mis-report/all-cust-arrear.model.ts
export interface AllCustomerArrearModel {
    loc: string;
    office: string;
    bg: string;
    bk: string;
    conNo: string;
    wlkOr: string;
    pvAc: string;
    name: string;
    fatherName: string;
    addr: string;
    likely: string;
    tarrif: string;
    nom: number;
    principalArrear: number;
    lpsArrear: number;
    vatArrear: number;
    totalArrear: number;
    billCycleCode: string;
    discDate: string;
    status: string;
}